consumidor-irregularidade/requirements.txt
aio-pika
python-dotenv
supabase